make scanning more efficient by skipping existing

+++ gallery/helpers.py
@@ -1,5 +1,6 @@
 from json import dumps, loads
 from os import makedirs
+from os.path import exists
 
 import requests
 from web3.auto import w3
@@ -32,8 +33,9 @@ def store_json(contract_address: str, token_id: int, data: dict):
     _p = f'{config.DATA_PATH}/{contract_address}'
     p = f'{_p}/{str(token_id)}.json'
     makedirs(_p, exist_ok=True)
-    with open(p, 'w') as f:
-        f.write(dumps(data))
+    if not exists(p):
+        with open(p, 'w') as f:
+            f.write(dumps(data))
 
 class Etherscan:
     def __init__(self, contract_address):
